Big big performance optim Background view are only loaded on demand
WARNING ! Hack with minimumHeight to fix the 0px height pb on first load.
on first load of the list view, as the views are 0px high, the listview generates as many views than they are items in the list, making the listview concept kind of useless and extremly slow to load.